command execution雛形作成
